requirements.txt
pandas==2.3.2
sqlalchemy==2.0.19
psycopg2-binary==2.9.7
python-dotenv==1.0.0